+++ src/ExponentialFamily.jl
@@ -38,20 +38,13 @@ include("distributions/dirichlet.jl")
 include("distributions/beta.jl")
 include("distributions/lognormal.jl")
 include("distributions/binomial.jl")
-# include("distributions/normal_mean_variance.jl")
-# include("distributions/normal_mean_precision.jl")
-# include("distributions/normal_weighted_mean_precision.jl")
 # include("distributions/multinomial.jl")
-# include("distributions/mv_normal_mean_covariance.jl")
-# include("distributions/mv_normal_mean_precision.jl")
-# include("distributions/mv_normal_weighted_mean_precision.jl")
-# include("distributions/normal.jl")
 # include("distributions/wishart.jl")
 # include("distributions/wishart_inverse.jl")
 # include("distributions/contingency.jl")
 include("distributions/erlang.jl")
 include("distributions/exponential.jl")
-# include("distributions/von_mises_fisher.jl")
+include("distributions/von_mises_fisher.jl")
 include("distributions/von_mises.jl")
 include("distributions/pareto.jl")
 # include("distributions/continuous_bernoulli.jl")
